refactor: rename Button component and start refactoring
Update component name for consistency
Begin restructuring for cleaner implementation
